ai_backend/requirements.txt
# Core API
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Image/Video Processing
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.26.2

# Face Recognition (existing, still used as fallback)
face_recognition==1.3.0

# Person Detection (YOLO)
ultralytics>=8.0.0

# Machine Learning
scikit-learn>=1.3.0

# Vector Database (lightweight, embedded)
chromadb>=0.4.0

# Deep Learning (optional - for OSNet Re-ID)
# Uncomment if using full OSNet:
# torch>=2.0.0
# torchvision>=0.15.0
# torchreid>=0.2.5
